+++ src/Root.scss
@@ -1,6 +1,4 @@
-@use "style/Essential.scss";
-@use "style/PathTree.scss";
-@use "style/Modals.scss";
+@import "app/common.blocks/index";
 
 #root {
   overflow: hidden; // UX: Fix potential issue on device emulator
@@ -14,15 +12,15 @@
     box-sizing: border-box;
 
     &[data-layout="classic"] {
-      @import "style/ClassicLayout.scss";
+      @import "app/classic.blocks/index";
     }
 
     &[data-layout="exclusive"] {
-      @import "style/ExclusiveLayout.scss";
+      @import "app/exclusive.blocks/index";
     }
 
     &[data-layout="mobile"] {
-      @import "style/MobileLayout.scss";
+      @import "app/mobile.blocks/index";
     }
 
     &[data-theme="light"] {
@@ -74,47 +72,6 @@
         background-color: #3c3c3c;
       }
     }
-  }
-}
-
-.flex-editor-panel {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 8px;
-}
-
-.menu-item.MuiMenuItem-root {
-  padding-left: 4px !important;
-
-  &.Mui-disabled {
-    pointer-events: auto !important;
-
-    &:hover {
-      // UX: don't change background color on hover
-      background-color: transparent !important;
-    }
-
-    &:active {
-      // https://moshfeu.github.io/show-tooltip-on-pointer-events-none-element/
-      pointer-events: none !important;
-    }
-  }
-
-  svg.menu-item-done {
-    height: 16px;
-    margin-right: 4px;
-    position: relative;
-    top: -1px;
-  }
-
-  svg.menu-item-next {
-    height: 24px;
-    position: relative;
-    top: -1px;
-    right: -8px;
   }
 }
