+++ src/app/classic.blocks/ControlAccordion.scss
@@ -1,0 +1,10 @@
+// @use "./index";
+
+.control-accordion {
+  text-align: left;
+  margin-bottom: 24px;
+  position: sticky;
+  top: -4px;
+  z-index: 100;
+  padding-top: $hpg;
+}
