+++ src/app/classic.blocks/SpeedCanvasElement.scss
@@ -1,0 +1,8 @@
+#speed-panel {
+  width: auto;
+  height: $speed-panel-height-vh;
+  height: $speed-panel-height-svh;
+  padding: $hpg;
+  overflow: hidden;
+  margin-top: $pg;
+}
