+++ src/app/classic.blocks/_index.tsx
@@ -1,0 +1,1 @@
+export {};
