+++ src/app/common.blocks/AboutModal.scss
@@ -1,0 +1,18 @@
+#about-modal {
+  padding: 32px;
+  width: 512px;
+  max-width: 80%;
+  min-height: 96px;
+  outline: none !important;
+
+  > img {
+    margin: auto;
+    width: 128px;
+    display: block;
+  }
+
+  a {
+    display: inline-block;
+    margin: 0 8px;
+  }
+}
