+++ src/app/common.blocks/AssetManagerModal.scss
@@ -1,117 +1,3 @@
-.modal-backdrop {
-  margin: 0;
-  background-color: transparent !important;
-  backdrop-filter: blur(12px);
-  outline: none !important;
-}
-
-.modal-container {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  color: var(--text-primary-color);
-}
-
-#welcome-modal {
-  padding: 16px;
-  width: 768px;
-  max-width: 80%;
-  min-height: 96px;
-  outline: none !important;
-  height: 80%;
-  overflow-y: auto;
-
-  .logo {
-    margin: auto;
-    width: 128px;
-    display: block;
-  }
-}
-
-@supports not (-moz-appearance: none) {
-  #welcome-modal {
-    padding-right: 10px;
-  }
-}
-
-#about-modal {
-  padding: 32px;
-  width: 512px;
-  max-width: 80%;
-  min-height: 96px;
-  outline: none !important;
-
-  > img {
-    margin: auto;
-    width: 128px;
-    display: block;
-  }
-
-  a {
-    display: inline-block;
-    margin: 0 8px;
-  }
-}
-
-#preferences-modal {
-  padding: 16px;
-  width: 512px;
-  max-width: 80%;
-  min-height: 96px;
-  outline: none !important;
-
-  hr {
-    margin: 16px 0;
-  }
-
-  .title {
-    margin-top: 16px;
-    margin-bottom: 16px;
-  }
-
-  .title:first-child {
-    margin-top: 0;
-  }
-}
-
-#confirmation-modal {
-  padding: 16px;
-  min-width: 320px;
-  max-width: 80%;
-  min-height: 96px;
-  outline: none !important;
-
-  .description-box {
-    display: flex;
-
-    > * {
-      flex-grow: 1;
-      width: 0;
-    }
-  }
-
-  .input-box {
-    width: 100%;
-
-    > * {
-      width: 100%;
-    }
-  }
-
-  .button-box {
-    display: flex;
-    margin-top: 16px;
-    justify-content: flex-end;
-    align-items: flex-end;
-    gap: 8px;
-
-    *:focus {
-      background-color: rgba(255, 255, 255, 0.1);
-    }
-  }
-}
-
 #asset-manager-modal {
   padding: 16px;
   width: 768px;
