+++ src/app/common.blocks/ConfirmationModal.tsx
@@ -1,11 +1,13 @@
 import { Box, Button, Card, Typography } from "@mui/material";
 import React from "react";
-import { getAppStores } from "../core/MainApp";
+import { getAppStores } from "../../core/MainApp";
 import { makeAutoObservable, action, when, observable, reaction } from "mobx";
 import { observer } from "mobx-react-lite";
-import { useMobxStorage } from "../core/Hook";
-import { ObserverInput } from "../component/ObserverInput";
-import { Modal } from "../component/Modal";
+import { useMobxStorage } from "../../core/Hook";
+import { ObserverInput } from "../../component/ObserverInput";
+import { Modal } from "./Modal";
+
+import "./ConfirmationModal.scss";
 
 export interface ConfirmationButton {
   label: string;
@@ -152,7 +154,7 @@ export const ConfirmationModal = observer(() => {
 
   return (
     <Modal symbol={ConfirmationModalSymbol} onClose={action(onClick.bind(null, -1))}>
-      <Card id="confirmation-modal" className="modal-container" onKeyDown={action(onKeyDown)}>
+      <Card className="confirmation-modal modal-container" onKeyDown={action(onKeyDown)}>
         <Typography variant="h2" gutterBottom>
           {cfm.title}
         </Typography>
