+++ src/app/common.blocks/PreferencesModal.scss
@@ -1,0 +1,20 @@
+#preferences-modal {
+  padding: 16px;
+  width: 512px;
+  max-width: 80%;
+  min-height: 96px;
+  outline: none !important;
+
+  hr {
+    margin: 16px 0;
+  }
+
+  .title {
+    margin-top: 16px;
+    margin-bottom: 16px;
+  }
+
+  .title:first-child {
+    margin-top: 0;
+  }
+}
