+++ src/app/common.blocks/PreferencesModal.tsx
@@ -1,14 +1,16 @@
 import { Card, Divider, Typography } from "@mui/material";
 import { observer } from "mobx-react-lite";
-import { getAppStores } from "../core/MainApp";
-import { AppThemeType } from "./Theme";
-import { clamp } from "../core/Util";
-import { ObserverEnumSelect } from "../component/ObserverEnumSelect";
-import { ObserverCheckbox } from "../component/ObserverCheckbox";
-import { ObserverInput } from "../component/ObserverInput";
-import { Modal } from "../component/Modal";
-import { enqueueInfoSnackbar } from "./Notice";
-import { Logger } from "../core/Logger";
+import { getAppStores } from "../../core/MainApp";
+import { AppThemeType } from "../Theme";
+import { clamp } from "../../core/Util";
+import { ObserverEnumSelect } from "../../component/ObserverEnumSelect";
+import { ObserverCheckbox } from "../../component/ObserverCheckbox";
+import { ObserverInput } from "../../component/ObserverInput";
+import { Modal } from "./Modal";
+import { enqueueInfoSnackbar } from "../Notice";
+import { Logger } from "../../core/Logger";
+
+import "./PreferencesModal.scss";
 
 export const PreferencesModalSymbol = Symbol("PreferencesModalSymbol");
