+++ src/app/common.blocks/SpeedCanvasElement.tsx
@@ -1,21 +1,23 @@
 import { makeObservable, action, observable, reaction } from "mobx";
 import { observer } from "mobx-react-lite";
-import { Point, Path, Vector, SpeedKeyframe, KeyframePos, BentRateApplicationDirection } from "../core/Path";
+import { Point, Path, Vector, SpeedKeyframe, KeyframePos, BentRateApplicationDirection } from "../../core/Path";
 import Konva from "konva";
 import { Circle, Layer, Line, Rect, Stage, Text } from "react-konva";
 import React from "react";
-import { PathConfig } from "../format/Config";
-import { AddKeyframe, MoveKeyframe, RemoveKeyframe, UpdateProperties } from "../core/Command";
-import { getAppStores } from "../core/MainApp";
-import { KeyframeIndexing } from "../core/Calculation";
-import { GraphCanvasConverter, getClientXY } from "../core/Canvas";
+import { PathConfig } from "../../format/Config";
+import { AddKeyframe, MoveKeyframe, RemoveKeyframe, UpdateProperties } from "../../core/Command";
+import { getAppStores } from "../../core/MainApp";
+import { KeyframeIndexing } from "../../core/Calculation";
+import { GraphCanvasConverter, getClientXY } from "../../core/Canvas";
 import { Box } from "@mui/material";
 import { Instance } from "@popperjs/core";
-import { useMobxStorage, useTouchEvent, useWindowSize } from "../core/Hook";
-import { getAppThemeInfo } from "./Theme";
-import { TouchEventListener } from "../core/TouchEventListener";
-import { Label, Padding0Tooltip } from "../component/TooltipLabel";
-import { getFieldCanvasHalfHeight } from "../core/Util";
+import { useMobxStorage, useTouchEvent, useWindowSize } from "../../core/Hook";
+import { getAppThemeInfo } from "../Theme";
+import { TouchEventListener } from "../../core/TouchEventListener";
+import { Label, Padding0Tooltip } from "../../component/TooltipLabel";
+import { getFieldCanvasHalfHeight } from "../../core/Util";
+
+import "./SpeedCanvasElement.scss";
 
 const FONT_FAMILY = '-apple-system,system-ui,BlinkMacSystemFont,"Segoe UI",Roboto,"Helvetica Neue",Arial,sans-serif';
