+++ src/app/common.blocks/WelcomeForOthers.mdx
@@ -1,4 +1,4 @@
-import { ObserverCheckbox } from "../component/ObserverCheckbox";
+import { ObserverCheckbox } from "../../component/ObserverCheckbox";
 
 <img src="static/logo464.svg" alt="app logo" className="logo" />
