+++ src/app/common.blocks/_index.scss
@@ -93,3 +93,12 @@ th {
     }
   }
 }
+
+.flex-editor-panel {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: nowrap;
+  justify-content: flex-start;
+  align-items: center;
+  gap: 8px;
+}
