+++ src/app/common.blocks/_index.tsx
@@ -1,0 +1,1 @@
+export {};
