+++ src/app/mobile.blocks/Welcome.scss
@@ -1,0 +1,7 @@
+#welcome-modal {
+  width: 100% !important;
+  height: 100% !important;
+  max-width: 100% !important;
+  padding: 8px;
+  box-sizing: border-box;
+}
