+++ src/app/mobile.blocks/_index.scss
@@ -1,5 +1,3 @@
-@import "Essential.scss";
-
 #bottom-nav {
   background-color: var(--bg-paper-color);
   position: fixed;
@@ -74,18 +72,6 @@
   }
 }
 
-#path-tree {
-  #path-tree-title {
-    position: relative;
-
-    > div {
-      position: absolute;
-      right: -8px;
-      top: -8px;
-    }
-  }
-}
-
 .panel-icon {
   width: 42px;
   height: 42px;
@@ -111,10 +97,5 @@
   }
 }
 
-#welcome-modal {
-  width: 100% !important;
-  height: 100% !important;
-  max-width: 100% !important;
-  padding: 8px;
-  box-sizing: border-box;
-}
+@import "./PathTreeAccordion";
+@import "./Welcome";
