+++ src/app/mobile.blocks/_index.tsx
@@ -1,0 +1,1 @@
+export {};
